Segregated into smaller components
